+++ kaithem/src/docs/changes.md
@@ -4,9 +4,13 @@ Change Log
 
 ### 0.84.0
 
+#### Added
+- :coffin: Fileserver resource types allow you to serve a directory of files as if it were at /pages/module/resourcename.  This replaces the old individual file permissions.
+
 #### Removed
 
 - :coffin: Web console removed due to lack of ASGI but may return later
+- :coffin: BREAKING. The internal fileref system.  Instead, files in modules are just simple files under \_\_filedata\_\_
 
 
 #### Changed